+++ java/src/advanced_java/lambdas/exercises/Main.java
@@ -1,6 +1,7 @@
 package advanced_java.lambdas.exercises;
 
 import advanced_java.lambdas.exercises.domain.Person;
+import advanced_java.lambdas.exercises.service.OperationService;
 
 import java.util.*;
 import java.util.function.Function;
@@ -51,7 +52,7 @@ public class Main {
         Scanner scanner = new Scanner(System.in);
 
         while (true) {
-            System.out.println("Escolha uma opção:");
+            System.out.println("\nEscolha uma opção:");
             System.out.println("[1] Filtrar idades pares");
             System.out.println("[2] Dobrar cada elemento");
             System.out.println("[3] Verificar se todos os elementos são maiores 10 anos");
@@ -64,7 +65,19 @@ public class Main {
             System.out.println("[10] Combinar duas listas intercalando elementos");
             System.out.print("Digite o número da opção: ");
 
-            int option = scanner.nextInt();
+            int option = -1;
+            boolean validInput = false;
+
+            while (!validInput) {
+                try {
+                    option = scanner.nextInt();
+                    validInput = true;
+                } catch (InputMismatchException e) {
+                    System.err.println("Entrada inválida. Por favor, digite um número inteiro.");
+                    scanner.next();
+                }
+            }
+
             scanner.nextLine();
 
             switch (option) {
@@ -78,75 +91,8 @@ public class Main {
                 case 8 -> System.out.println("Média das idades: " + service.calculateAverage(ages));
                 case 9 -> System.out.println("Idades distintas: " + service.removeDuplicates(ages));
                 case 10 -> System.out.println("Listas combinadas: " + service.combineListsInterleaving(names, ages));
-                default -> System.out.println("Opção inválida.");
+                default -> System.err.println("Opção inválida.");
             }
-        }
-    }
-
-    public static class OperationService {
-
-        public List<Integer> filterEvenNumbers(List<Integer> numbers) {
-            return numbers.stream()
-                    .filter(n -> n % 2 == 0)
-                    .collect(Collectors.toList());
-        }
-
-        public List<Integer> doubleElements(List<Integer> numbers) {
-            return numbers.stream()
-                    .map(n -> n * 2)
-                    .collect(Collectors.toList());
-        }
-
-        public boolean areAllElementsGreaterThan(List<Integer> numbers, int value) {
-            return numbers.stream()
-                    .allMatch(n -> n > value);
-        }
-
-        public Optional<Integer> findMax(List<Integer> numbers) {
-            return numbers.stream()
-                    .max(Integer::compareTo);
-        }
-
-        public List<String> convertToUpperCase(List<String> strings) {
-            return strings.stream()
-                    .map(String::toUpperCase)
-                    .collect(Collectors.toList());
-        }
-
-        public List<String> sortByLength(List<String> strings) {
-            return strings.stream()
-                    .sorted(Comparator.comparingInt(String::length))
-                    .collect(Collectors.toList());
-        }
-
-        public Map<Integer, List<Person>> groupPeopleByAge(List<Person> people) {
-            return people.stream()
-                    .collect(Collectors.groupingBy(Person::getAge));
-        }
-
-        public double calculateAverage(List<Integer> numbers) {
-            return numbers.stream()
-                    .collect(Collectors.averagingDouble(Integer::doubleValue));
-        }
-
-        public List<Integer> removeDuplicates(List<Integer> numbers) {
-            return numbers.stream()
-                    .distinct()
-                    .collect(Collectors.toList());
-        }
-
-        public <T> List<T> combineListsInterleaving(List<String> list1, List<Integer> list2) {
-            List<T> combined = new ArrayList<>();
-            int maxSize = Math.max(list1.size(), list2.size());
-            for (int i = 0; i < maxSize; i++) {
-                if (i < list1.size()) {
-                    combined.add((T) list1.get(i));
-                }
-                if (i < list2.size()) {
-                    combined.add((T) list2.get(i));
-                }
-            }
-            return combined;
         }
     }
 }
